move more shared things to the base
